+++ oidc_provider/migrations/0028_change_response_types_field_1_of_3.py
@@ -1,0 +1,21 @@
+from django.db import migrations
+from django.db import models
+
+
+class Migration(migrations.Migration):
+    dependencies = [
+        ("oidc_provider", "0027_alter_rsakey_options"),
+    ]
+
+    operations = [
+        migrations.RenameField(
+            model_name="client",
+            old_name="response_types",
+            new_name="old_response_types",
+        ),
+        migrations.AddField(
+            model_name="client",
+            name="response_types",
+            field=models.JSONField(default=list),
+        ),
+    ]