+++ oidc_provider/models.py
@@ -10,6 +10,8 @@ from django.db import models
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
+from oidc_provider.fields import JSONMultiSelectModelField
+
 CLIENT_TYPE_CHOICES = [
     ("confidential", "Confidential"),
     ("public", "Public"),
@@ -28,31 +30,6 @@ JWT_ALGS = [
     ("HS256", "HS256"),
     ("RS256", "RS256"),
 ]
-
-
-class ResponseTypeManager(models.Manager):
-    def get_by_natural_key(self, value):
-        return self.get(value=value)
-
-
-class ResponseType(models.Model):
-    objects = ResponseTypeManager()
-
-    value = models.CharField(
-        max_length=30,
-        choices=RESPONSE_TYPE_CHOICES,
-        unique=True,
-        verbose_name=_("Response Type Value"),
-    )
-    description = models.CharField(
-        max_length=50,
-    )
-
-    def natural_key(self):
-        return (self.value,)  # natural_key must return tuple
-
-    def __str__(self):
-        return "{0}".format(self.description)
 
 
 class Client(models.Model):
@@ -78,7 +55,9 @@ class Client(models.Model):
     )
     client_id = models.CharField(max_length=255, unique=True, verbose_name=_("Client ID"))
     client_secret = models.CharField(max_length=255, blank=True, verbose_name=_("Client SECRET"))
-    response_types = models.ManyToManyField(ResponseType)
+    response_types = JSONMultiSelectModelField(
+        choices=RESPONSE_TYPE_CHOICES, verbose_name=_("Response Types")
+    )
     jwt_alg = models.CharField(
         max_length=10,
         choices=JWT_ALGS,
@@ -143,11 +122,13 @@ class Client(models.Model):
         return self.__str__()
 
     def response_type_values(self):
-        return (response_type.value for response_type in self.response_types.all())
+        # Return the allowed response types in the same order as they appear in
+        # RESPONSE_TYPE_CHOICES.
+        return [code for code, description in RESPONSE_TYPE_CHOICES if code in self.response_types]
 
     def response_type_descriptions(self):
-        # return as a list, rather than a generator, so descriptions display correctly in admin
-        return [response_type.description for response_type in self.response_types.all()]
+        response_type_dict = dict(RESPONSE_TYPE_CHOICES)
+        return [response_type_dict[response_type] for response_type in self.response_type_values()]
 
     @property
     def redirect_uris(self):