+++ oidc_provider/tests/app/utils.py
@@ -17,7 +17,6 @@ from django.utils import timezone
 from oidc_provider.lib.claims import ScopeClaims
 from oidc_provider.models import Client
 from oidc_provider.models import Code
-from oidc_provider.models import ResponseType
 from oidc_provider.models import Token
 
 FAKE_NONCE = "cb584e44c43ed6bd0bc2d9c7e242837d"
@@ -71,8 +70,8 @@ def create_fake_client(response_type, is_public=False, require_consent=True):
     if isinstance(response_type, ("".__class__, "".__class__)):
         response_type = (response_type,)
     for value in response_type:
-        client.response_types.add(ResponseType.objects.get(value=value))
-
+        client.response_types.add(value)
+    client.save()
     return client