+++ oidc_provider/tests/cases/test_authorize_endpoint.py
@@ -12,6 +12,7 @@ try:
 except ImportError:
     from urlparse import parse_qs
     from urlparse import urlsplit
+
 import uuid
 from unittest.mock import Mock
 from unittest.mock import patch
@@ -22,6 +23,7 @@ try:
     from django.urls import reverse
 except ImportError:
     from django.core.urlresolvers import reverse
+
 from django.contrib.auth.models import AnonymousUser
 from django.core.management import call_command
 from django.test import RequestFactory
@@ -385,7 +387,7 @@ class AuthorizationCodeFlowTestCase(TestCase, AuthorizeEndpointMixin):
         """
         data = {
             "client_id": self.client_code.client_id,
-            "response_type": next(self.client_code.response_type_values()),
+            "response_type": self.client_code.response_type_values()[0],
             "redirect_uri": self.client_code.default_redirect_uri,
             "scope": "openid email",
             "state": self.state,
@@ -412,7 +414,7 @@ class AuthorizationCodeFlowTestCase(TestCase, AuthorizeEndpointMixin):
         """
         data = {
             "client_id": self.client_code.client_id,
-            "response_type": next(self.client_code.response_type_values()),
+            "response_type": self.client_code.response_type_values()[0],
             "redirect_uri": self.client_code.default_redirect_uri,
             "scope": "openid email",
             "state": self.state,
@@ -444,7 +446,7 @@ class AuthorizationCodeFlowTestCase(TestCase, AuthorizeEndpointMixin):
         """
         data = {
             "client_id": self.client_code.client_id,
-            "response_type": next(self.client_code.response_type_values()),
+            "response_type": self.client_code.response_type_values()[0],
             "redirect_uri": self.client_code.default_redirect_uri,
             "scope": "openid email",
             "state": self.state,
@@ -466,7 +468,7 @@ class AuthorizationCodeFlowTestCase(TestCase, AuthorizeEndpointMixin):
         """
         data = {
             "client_id": self.client_code.client_id,
-            "response_type": next(self.client_code.response_type_values()),
+            "response_type": self.client_code.response_type_values()[0],
             "redirect_uri": self.client_code.default_redirect_uri,
             "scope": "openid email",
             "state": self.state,
@@ -488,7 +490,7 @@ class AuthorizationCodeFlowTestCase(TestCase, AuthorizeEndpointMixin):
         """
         data = {
             "client_id": self.client_code.client_id,
-            "response_type": next(self.client_code.response_type_values()),
+            "response_type": self.client_code.response_type_values()[0],
             "redirect_uri": self.client_code.default_redirect_uri,
             "scope": "openid email",
             "state": self.state,
@@ -516,7 +518,7 @@ class AuthorizationCodeFlowTestCase(TestCase, AuthorizeEndpointMixin):
 
         data = {
             "client_id": self.client_code.client_id,
-            "response_type": next(self.client_code.response_type_values()),
+            "response_type": self.client_code.response_type_values()[0],
             "redirect_uri": self.client_code.default_redirect_uri,
             "scope": "openid email",
             "state": self.state,
@@ -541,7 +543,7 @@ class AuthorizationCodeFlowTestCase(TestCase, AuthorizeEndpointMixin):
 
         data = {
             "client_id": self.client_code.client_id,
-            "response_type": next(self.client_code.response_type_values()),
+            "response_type": self.client_code.response_type_values()[0],
             "redirect_uri": self.client_code.default_redirect_uri,
             "scope": "openid email",
             "state": self.state,
@@ -612,7 +614,7 @@ class AuthorizationImplicitFlowTestCase(TestCase, AuthorizeEndpointMixin):
         """
         data = {
             "client_id": self.client_code.client_id,
-            "response_type": next(self.client_code.response_type_values()),
+            "response_type": self.client_code.response_type_values()[0],
             "redirect_uri": self.client_code.default_redirect_uri,
             "scope": "openid email",
             "state": self.state,
@@ -630,7 +632,7 @@ class AuthorizationImplicitFlowTestCase(TestCase, AuthorizeEndpointMixin):
         data = {
             "client_id": self.client_code.client_id,
             "redirect_uri": self.client_code.default_redirect_uri,
-            "response_type": next(self.client_code.response_type_values()),
+            "response_type": self.client_code.response_type_values()[0],
             "scope": "openid email",
             "state": self.state,
             "nonce": self.nonce,
@@ -645,7 +647,7 @@ class AuthorizationImplicitFlowTestCase(TestCase, AuthorizeEndpointMixin):
         # same for public client
         data["client_id"] = (self.client_public.client_id,)
         data["redirect_uri"] = (self.client_public.default_redirect_uri,)
-        data["response_type"] = (next(self.client_public.response_type_values()),)
+        data["response_type"] = (self.client_public.response_type_values()[0],)
 
         response = self._auth_request("post", data, is_user_authenticated=True)
 
@@ -660,7 +662,7 @@ class AuthorizationImplicitFlowTestCase(TestCase, AuthorizeEndpointMixin):
         data = {
             "client_id": self.client_no_access.client_id,
             "redirect_uri": self.client_no_access.default_redirect_uri,
-            "response_type": next(self.client_no_access.response_type_values()),
+            "response_type": self.client_no_access.response_type_values()[0],
             "scope": "openid email",
             "state": self.state,
             "nonce": self.nonce,
@@ -675,7 +677,7 @@ class AuthorizationImplicitFlowTestCase(TestCase, AuthorizeEndpointMixin):
         # same for public client
         data["client_id"] = (self.client_public_no_access.client_id,)
         data["redirect_uri"] = (self.client_public_no_access.default_redirect_uri,)
-        data["response_type"] = (next(self.client_public_no_access.response_type_values()),)
+        data["response_type"] = (self.client_public_no_access.response_type_values()[0],)
 
         response = self._auth_request("post", data, is_user_authenticated=True)
 
@@ -690,7 +692,7 @@ class AuthorizationImplicitFlowTestCase(TestCase, AuthorizeEndpointMixin):
         data = {
             "client_id": self.client_code.client_id,
             "redirect_uri": self.client_code.default_redirect_uri,
-            "response_type": next(self.client_code.response_type_values()),
+            "response_type": self.client_code.response_type_values()[0],
             "scope": "openid email",
             "state": self.state,
             "nonce": self.nonce,
@@ -716,7 +718,7 @@ class AuthorizationImplicitFlowTestCase(TestCase, AuthorizeEndpointMixin):
         data = {
             "client_id": self.client_no_access.client_id,
             "redirect_uri": self.client_no_access.default_redirect_uri,
-            "response_type": next(self.client_no_access.response_type_values()),
+            "response_type": self.client_no_access.response_type_values()[0],
             "scope": "openid email",
             "state": self.state,
             "nonce": self.nonce,
@@ -740,7 +742,7 @@ class AuthorizationImplicitFlowTestCase(TestCase, AuthorizeEndpointMixin):
         """
         data = {
             "client_id": self.client_public_no_consent.client_id,
-            "response_type": next(self.client_public_no_consent.response_type_values()),
+            "response_type": self.client_public_no_consent.response_type_values()[0],
             "redirect_uri": self.client_public_no_consent.default_redirect_uri,
             "scope": "openid email",
             "state": self.state,
@@ -803,7 +805,7 @@ class AuthorizationHybridFlowTestCase(TestCase, AuthorizeEndpointMixin):
         self.data = {
             "client_id": self.client_code_idtoken_token.client_id,
             "redirect_uri": self.client_code_idtoken_token.default_redirect_uri,
-            "response_type": next(self.client_code_idtoken_token.response_type_values()),
+            "response_type": self.client_code_idtoken_token.response_type_values()[0],
             "scope": "openid email",
             "state": self.state,
             "nonce": self.nonce,
@@ -849,7 +851,7 @@ class TestCreateResponseURI(TestCase):
         data = {
             "client_id": client.client_id,
             "redirect_uri": client.default_redirect_uri,
-            "response_type": next(client.response_type_values()),
+            "response_type": client.response_type_values()[0],
         }
 
         factory = RequestFactory()